# backend/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI-Library-Backend-API",
    "description": "Backend API Document"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/api/ai-assistant/sessions": {
      "get": {
        "tags": ["AI-Assistant"],
        "summary": "获取会话列表",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "会话列表"
          }
        }
      },
      "post": {
        "tags": ["AI-Assistant"],
        "summary": "创建新的聊天会话",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  }
                },
                "example": {
                  "title": "AI Service Test Session"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功"
          }
        }
      }
    },
    "/api/ai-assistant/sessions/{id}": {
      "patch": {
        "tags": ["AI-Assistant"],
        "summary": "更新会话标题",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  }
                },
                "example": {
                  "title": "Updated Session Title"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          }
        }
      },
      "delete": {
        "tags": ["AI-Assistant"],
        "summary": "删除会话",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        }
      }
    },
    "/api/ai-assistant/sessions/{id}/messages": {
      "get": {
        "tags": ["AI-Assistant"],
        "summary": "获取消息历史",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "消息列表"
          }
        }
      },
      "post": {
        "tags": ["AI-Assistant"],
        "summary": "发送消息",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string"
                  },
                  "stream": {
                    "type": "boolean"
                  }
                },
                "example": {
                  "content": "你好",
                  "stream": false
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "发送成功"
          }
        }
      }
    },
    "/api/ai-assistant/sessions/{id}/system-messages": {
      "post": {
        "tags": ["AI-Assistant"],
        "summary": "保存系统消息",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "metadata": {
                    "type": "object"
                  }
                },
                "example": {
                  "role": "system",
                  "content": "你是一个乐于助人的图书馆助手",
                  "metadata": {
                    "intent": "role_setting"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "保存成功"
          }
        }
      }
    },
    "/api/ai-assistant/sessions/{id}/messages/stream": {
      "get": {
        "tags": ["AI-Assistant"],
        "summary": "流式请求消息",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "message",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "流式响应"
          }
        }
      }
    },
    "/api/ai-assistant/messages/{id}/feedback": {
      "post": {
        "tags": ["AI-Assistant"],
        "summary": "提交消息反馈",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "feedback_type": {
                    "type": "string"
                  },
                  "feedback_reason": {
                    "type": "string"
                  },
                  "feedback_comment": {
                    "type": "string"
                  }
                },
                "example": {
                  "feedback_type": "like",
                  "feedback_reason": "helpful",
                  "feedback_comment": "回答准确"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "提交成功"
          }
        }
      }
    },
...
  }
}